S3-T14: CI release gate with DB guard, full suites, and Playwright smoke (#85)
DB-backed tests import a shared guard that hard-fails unless DATABASE_URL
is localhost/clipmind_dev or CLIPMIND_TEST_DB_OVERRIDE=1 (CI service
containers only). Push tests scope to fixture creators via a creatorIds
test filter instead of disabling every other creator's push. CI now runs
six gating jobs with no continue-on-error: env-file hygiene, full
typecheck, all pure Node suites (108), DB suites (85) against postgres:17
with migrate deploy, the Playwright smoke (build + seed + login + Home/
Review/Rhythm render), and clip-service ruff + pytest. E2E deliberately
excludes upload/clip-service/Minds/OpenAI paths - documented in README.

Verification fixes: strict-mode locator in the smoke (2 seeded clips),
and 10 ruff findings in clip-service (RUF046 casts, import sort, G201
logger.exception) so the new ruff gate starts green.

Local proof: test:db 85/85, test:node 108/108, test:e2e 1/1 passed,
pytest 43/43, ruff all checks passed, typecheck clean.

+++ clip-service/src/transcribe.py
@@ -68,7 +68,7 @@ def probe_video_duration_ms(video_path: Path | str) -> int:
     if duration_seconds is None:
         raise RuntimeError("ffprobe did not return a video duration.")
 
-    return max(0, int(round(float(duration_seconds) * 1000)))
+    return max(0, round(float(duration_seconds) * 1000))
 
 
 def extract_audio(video_path: Path, audio_path: Path) -> None:
@@ -192,4 +192,4 @@ def _response_to_dict(response: Any) -> dict[str, Any]:
 def _seconds_to_ms(value: Any) -> int:
     if value is None:
         return 0
-    return max(0, int(round(float(value) * 1000)))
+    return max(0, round(float(value) * 1000))
